feat(cosign): use sigstore to verify the hashdir cli
Not only do we sign the bootstrap script it's self with cosign
in our release pipeline but we now also use it to verify the
hashdir installation at runtime, with more use cases to follow.

Also the hashdir cli and cosign it's self are now installed into a
"utils" folder separate from the "bin" folder that we use for our
shims. This is so that the script's version of these tools do not
conflict with the same tool that the user may have installed for
use in their own project.